change(Popper): Move component
